Add upload size to upload button
